vhal-gen/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vhal-gen"
version = "1.0.0"
description = "FLYNC YAML to Android VHAL code generator"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "streamlit>=1.28",
]

[project.scripts]
vhal-gen = "vhal_gen.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
